+++ examples/flag_and_option_specifications.rb
@@ -21,7 +21,7 @@ InfoLines = [
 # Specify specifications, parse, and checking standard flags
 
 Flag_Debug = CLASP.Flag('--debug', alias: '-d', help: 'runs in Debug mode')
-Option_Verbosity = CLASP.Option('--verbosity', alias: '-v', help: 'specifies the verbosity', values: [ 'terse', 'quiet', 'silent', 'chatty' ])
+Option_Verbosity = CLASP.Option('--verbosity', alias: '-v', help: 'specifies the verbosity', values: [ 'terse', 'quiet', 'silent', 'chatty' ], default_value: 'terse')
 Flag_Chatty = CLASP.Flag('--verbosity=chatty', alias: '-c')
 
 Specifications = [
@@ -38,7 +38,7 @@ args = CLASP::Arguments.new ARGV, Specifications
 
 if args.flags.include?(CLASP::FlagSpecification.Help)
 
-    CLASP.show_usage(Specifications, exit_code: 0, version: ProgramVersion, stream: $stdout, info_lines: InfoLines)
+    CLASP.show_usage(Specifications, exit_code: 0, version: ProgramVersion, stream: $stdout, info_lines: InfoLines, default_indicator: '*default*')
 end
 
 if args.flags.include?('--version')